Update!
I lost the original source, yes I know it was on GitHub. Discord.Net update a hella bunch so I just redid the whole thing.
