+++ src/LDMAuthenticator.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LDMAuthenticator", "LDMAuthenticator\LDMAuthenticator.csproj", "{0A6B8AC0-F25C-4C6C-A4CA-C86DCA67EC87}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{0A6B8AC0-F25C-4C6C-A4CA-C86DCA67EC87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0A6B8AC0-F25C-4C6C-A4CA-C86DCA67EC87}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0A6B8AC0-F25C-4C6C-A4CA-C86DCA67EC87}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0A6B8AC0-F25C-4C6C-A4CA-C86DCA67EC87}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
