+++ src/LDMAuthenticator/Enum/ELogType.cs
@@ -1,0 +1,8 @@
+﻿namespace LDMAuthenticator.Enum;
+
+public enum ELogType
+{
+    Debug,
+    Info,
+    Error
+}
