+++ src/LDMAuthenticator/configuration.yaml
@@ -1,0 +1,2 @@
+﻿client:
+  token:
