--- src/LDMBot.sln
@@ -1,16 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LDMBot", "LDMBot\LDMBot.csproj", "{ACF0282F-2790-484C-AA64-CAFCFE46FE01}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{ACF0282F-2790-484C-AA64-CAFCFE46FE01}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ACF0282F-2790-484C-AA64-CAFCFE46FE01}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ACF0282F-2790-484C-AA64-CAFCFE46FE01}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ACF0282F-2790-484C-AA64-CAFCFE46FE01}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
