--- src/LDMBot/Program.cs
@@ -1,10 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace papershredder432.LDMAuthenticator
-{
-    class Program
-    {
-        public static Task Main(string[] args)
-            => Startup.RunAsync(args);
-    }
-}
