--- src/LDMBot/packages.config
@@ -1,30 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Discord.Net" version="2.2.0" targetFramework="net472" />
-  <package id="Discord.Net.Commands" version="2.2.0" targetFramework="net472" />
-  <package id="Discord.Net.Core" version="2.2.0" targetFramework="net472" />
-  <package id="Discord.Net.Rest" version="2.2.0" targetFramework="net472" />
-  <package id="Discord.Net.WebSocket" version="2.2.0" targetFramework="net472" />
-  <package id="Discord.Net.Webhook" version="2.2.0" targetFramework="net472" />
-  <package id="Microsoft.Bcl.AsyncInterfaces" version="1.1.1" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration" version="3.1.7" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration.Abstractions" version="3.1.7" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Configuration.FileExtensions" version="2.0.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.DependencyInjection" version="3.1.7" targetFramework="net472" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net472" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="3.1.7" targetFramework="net472" />
-  <package id="Microsoft.Extensions.FileProviders.Abstractions" version="2.0.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.FileProviders.Physical" version="2.0.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.FileSystemGlobbing" version="2.0.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Primitives" version="3.1.7" targetFramework="net472" />
-  <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net472" />
-  <package id="System.Memory" version="4.5.2" targetFramework="net472" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.7.1" targetFramework="net472" />
-  <package id="System.Buffers" version="4.4.0" targetFramework="net472" />
-  <package id="System.Collections.Immutable" version="1.3.1" targetFramework="net472" />
-  <package id="System.Interactive.Async" version="4.0.0" targetFramework="net472" />
-  <package id="System.Linq.Async" version="4.0.0" targetFramework="net472" />
-  <package id="System.Numerics.Vectors" version="4.4.0" targetFramework="net472" />
-  <package id="Unofficial.Microsoft.Extensions.Configuration.Yaml" version="19.232.0.38254" targetFramework="net472" />
-  <package id="YamlDotNet" version="6.1.2" targetFramework="net472" />
-</packages>
